+++ vm/src/hint_processor/builtin_hint_processor/secp/field_utils.rs
@@ -196,7 +196,6 @@ mod tests {
     use super::*;
     use crate::hint_processor::builtin_hint_processor::hint_code;
     use crate::stdlib::string::ToString;
-    use crate::vm::runners::cairo_runner::RunResources;
 
     use crate::{
         any_box,
@@ -204,7 +203,7 @@ mod tests {
             builtin_hint_processor::builtin_hint_processor_definition::{
                 BuiltinHintProcessor, HintProcessorData,
             },
-            hint_processor_definition::HintProcessor,
+            hint_processor_definition::HintProcessorLogic,
         },
         types::{
             exec_scope::ExecutionScopes,