+++ vm/src/hint_processor/builtin_hint_processor/secp/signature.rs
@@ -165,7 +165,6 @@ mod tests {
     use crate::stdlib::string::ToString;
     use crate::types::errors::math_errors::MathError;
 
-    use crate::vm::runners::cairo_runner::RunResources;
     use crate::{
         any_box,
         hint_processor::{
@@ -173,7 +172,7 @@ mod tests {
                 builtin_hint_processor_definition::{BuiltinHintProcessor, HintProcessorData},
                 hint_code,
             },
-            hint_processor_definition::HintProcessor,
+            hint_processor_definition::HintProcessorLogic,
         },
         types::exec_scope::ExecutionScopes,
         utils::test_utils::*,