+++ vm/src/utils.rs
@@ -427,13 +427,7 @@ pub mod test_utils {
         ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr, $constants:expr) => {{
             let hint_data = HintProcessorData::new_default($hint_code.to_string(), $ids_data);
             let mut hint_processor = BuiltinHintProcessor::new_empty();
-            hint_processor.execute_hint(
-                &mut $vm,
-                $exec_scopes,
-                &any_box!(hint_data),
-                $constants,
-                &mut RunResources::default(),
-            )
+            hint_processor.execute_hint(&mut $vm, $exec_scopes, &any_box!(hint_data), $constants)
         }};
         ($vm:expr, $ids_data:expr, $hint_code:expr, $exec_scopes:expr) => {{
             let hint_data = HintProcessorData::new_default(
@@ -446,7 +440,6 @@ pub mod test_utils {
                 $exec_scopes,
                 &any_box!(hint_data),
                 &crate::stdlib::collections::HashMap::new(),
-                &mut RunResources::default(),
             )
         }};
         ($vm:expr, $ids_data:expr, $hint_code:expr) => {{
@@ -460,7 +453,6 @@ pub mod test_utils {
                 exec_scopes_ref!(),
                 &any_box!(hint_data),
                 &crate::stdlib::collections::HashMap::new(),
-                &mut RunResources::default(),
             )
         }};
     }
@@ -605,15 +597,15 @@ pub mod test_utils {
 
 #[cfg(test)]
 mod test {
+    use crate::hint_processor::hint_processor_definition::HintProcessorLogic;
     use crate::stdlib::{cell::RefCell, collections::HashMap, rc::Rc, string::String, vec::Vec};
-    use crate::vm::runners::cairo_runner::RunResources;
     use crate::{
         hint_processor::{
             builtin_hint_processor::{
                 builtin_hint_processor_definition::{BuiltinHintProcessor, HintProcessorData},
                 dict_manager::{DictManager, DictTracker},
             },
-            hint_processor_definition::{HintProcessor, HintReference},
+            hint_processor_definition::HintReference,
         },
         serde::deserialize_program::{BuiltinName, ReferenceManager},
         types::{exec_scope::ExecutionScopes, program::Program, relocatable::MaybeRelocatable},